Fix item stuck in workqueue after trigger renames (#98)
Items still in the workqueue, mostly due to template errors, were able to be stuck forever in the queue as the controller repeatedly tried to retrieve the cache backing the trigger.

If any such item is retrieved the item is now passed without the triggering resources, same as if the resource itself got deleted.