Refactor the dynamic controllers watches
This pull request changes how the watches and the cache is built.

With this PR the caches and watches are set up before the dynamic controllers are started.
Building the caches is moved to the manager of the dynamic controllers.
If caches need to be reconfigured the manager shuts down the controller and builds and starts a new one.
This removes the need for a lot of synchronization and should lead to less edge-cases when reconfiguring a managed resource.